refactor(adapters): make the adapter set an open registry, not a central map
Replace the single `_KNOWN` mapping with a registry: each warehouse is a
self-contained module under `adapters/builtin/` that builds an `AdapterProfile`
and calls `register(...)`, and the registry auto-discovers those modules on first
lookup. Adding a warehouse is now a new file and nothing else, with no central
map to edit; an out-of-tree package extends support the same way by calling
`register` at import.

`adapters.py` becomes a package: `model` holds the value types, `registry` holds
registration, discovery, and resolution, and `builtin/` holds one module per
adapter. `VALIDATED_ADAPTERS` (a constant computed from the central map) becomes
`validated_adapters()` since the set is now whatever has registered.
