+++ docs/current_state/architecture.md
@@ -115,13 +115,15 @@ Three modules:
 - [**`patterns.py`**](../../src/dblect/sql/patterns.py): list queries and detectors over the AST.
 - [**`_sqlglot.py`**](../../src/dblect/sql/_sqlglot.py): typed accessors over sqlglot's `Any`-heavy attribute surface.
 
-### The target adapter (`src/dblect/adapters.py`)
+### The target adapter (`src/dblect/adapters/`)
 
-A dbt project compiles against one adapter (duckdb, snowflake, bigquery, ...), and that single choice fixes everything dblect reasons about the target: which sqlglot dialect parses its compiled SQL, whether the warehouse enforces `PRIMARY KEY` / `UNIQUE` and `NOT NULL` on write, and which incremental strategy runs when a model leaves `incremental_strategy` unset. [`AdapterProfile`](../../src/dblect/adapters.py) gathers those facets into one value so a run reads a single coherent target rather than assembling it from independent per-facet lookups. A dbt adapter name and a sqlglot dialect name are two namespaces that overlap by name without being the same thing; the profile carries both.
+A dbt project compiles against one adapter (duckdb, snowflake, bigquery, ...), and that single choice fixes everything dblect reasons about the target: which sqlglot dialect parses its compiled SQL, whether the warehouse enforces `PRIMARY KEY` / `UNIQUE` and `NOT NULL` on write, and which incremental strategy runs when a model leaves `incremental_strategy` unset. [`AdapterProfile`](../../src/dblect/adapters/model.py) gathers those facets into one value so a run reads a single coherent target rather than assembling it from independent per-facet lookups. A dbt adapter name and a sqlglot dialect name are two namespaces that overlap by name without being the same thing; the profile carries both.
 
-`profile_for_adapter(adapter_type)` is the semantics lookup: it returns the profile for a known adapter and a conservative profile for any other, never raising. `resolve_profile(adapter_type, explicit_dialect)` is the parsing-validation gate the CLI calls after loading the manifest. An explicit `--dialect` names the target wholesale, so its grammar and its runtime semantics always agree (forcing `snowflake` gives snowflake's dialect *and* snowflake's enforcement, never a hybrid); passing the flag is the operator's acknowledgment of a best-effort interpretation. Without it, the manifest's adapter must be validated, otherwise `UnvalidatedAdapterError` fires and the CLI bails with a message that names the adapter, the validated set, and the `--dialect` escape. When the resolved target is not validated, the CLI prints a one-line stderr warning so the run is never silently best-effort.
+Profiles live behind a [registry](../../src/dblect/adapters/registry.py): each warehouse is a self-contained module under [`adapters/builtin/`](../../src/dblect/adapters/builtin/) that builds an `AdapterProfile` and calls `register(...)`. The registry auto-discovers those modules on first lookup, so **adding a warehouse is a new file and nothing else** (no central map to edit); an out-of-tree package extends support the same way, by calling `register` at import.
 
-An adapter is **validated** when dblect's detectors have been exercised against its SQL end-to-end. Only `duckdb` is validated today (`VALIDATED_ADAPTERS`).
+`profile_for_adapter(adapter_type)` is the semantics lookup: it returns the registered profile for a known adapter and a conservative profile for any other, never raising. `resolve_profile(adapter_type, explicit_dialect)` is the parsing-validation gate the CLI calls after loading the manifest. An explicit `--dialect` names the target wholesale, so its grammar and its runtime semantics always agree (forcing `snowflake` gives snowflake's dialect *and* snowflake's enforcement, never a hybrid); passing the flag is the operator's acknowledgment of a best-effort interpretation. Without it, the manifest's adapter must be validated, otherwise `UnvalidatedAdapterError` fires and the CLI bails with a message that names the adapter, the validated set, and the `--dialect` escape. When the resolved target is not validated, the CLI prints a one-line stderr warning so the run is never silently best-effort.
+
+An adapter is **validated** when dblect's detectors have been exercised against its SQL end-to-end. Only `duckdb` is validated today (`validated_adapters()`).
 
 The resolved `AdapterProfile` is threaded as the single target through `run_audit`, `run_check`, the detector makers, and the property constructors; the leaf parsing functions take the profile's `sqlglot_dialect`. Programmatic callers pass a profile directly (e.g. `profile_for_adapter("duckdb")`); the validation gate lives at the CLI boundary, not in the analysis pipeline.
